+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-tbd
+### Fixed
+- Fix a bug where the plugin was not able to detect the frameworks for some project configurations
+[#8](https://github.com/ge-org/multiplatform-swiftpackage/issues/8)
 
 ## [2.0.1]
 ### Changed